refactor: split registrar_integration.py (3510 LOC → 7 modules)
Split god file into package:
- _helpers.py (373 LOC) — imports, constants, helper functions, router
- _departments.py (70 LOC) — departments list endpoint
- _queue_profiles.py (416 LOC) — queue profiles CRUD + reorder
- _services_doctors.py (268 LOC) — services + doctors endpoints
- _queue_ops.py (1394 LOC) — queue-settings, generate-qr,
  open-reception, start-visit, calendar, batch queue
- _today_queues.py (989 LOC) — today queues aggregation,
  doctor user-id conversion
- __init__.py (20 LOC) — re-exports router + get_today_queues +
  start_queue_visit (used by lab_reporting.py and registrar_wizard)

External imports from 2 files maintained via __init__.py re-exports.

Tests: app loads OK; frontend 554/554 pass (0 regressions).
